% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteToken.R
\name{disconnectFromOpenSILEXWS}
\alias{disconnectFromOpenSILEXWS}
\title{disconnect a user identifier for connexion to the web service}
\usage{
disconnectFromOpenSILEXWS()
}
\value{
a boolean TRUE if succeeeded
                  FALSE if failed
}
\description{
disconnect a user identifier for connexion to the web service
}
\examples{
\donttest{
disconnectFromOpenSILEXWS()
}
}
