% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getDataAndMetadataFromResponse}
\alias{getDataAndMetadataFromResponse}
\title{getDataAndMetadataFromResponse}
\usage{
getDataAndMetadataFromResponse(responseObject)
}
\arguments{
\item{responseObject}{objet de reponse HTTP httr}
}
\description{
Retreive httr response status and data linked to it
}
