% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getOperations}
\alias{getOperations}
\title{getOperations}
\usage{
getOperations()
}
\value{
an object with raw informations about operations from opensilex api
}
\description{
show operations from opensilex api
}
