% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{retrieves a user identifier for connexion to the web service}
\usage{
getToken(login, password)
}
\arguments{
\item{login}{login of the user to create the token}

\item{password}{password of the user to create the token}
}
\value{
a session token user identifier in the WS
}
\description{
Retrieves a user identifier for connexion to the WebService (WS)
}
\examples{
\donttest{
connectToOpenSILEXWS("guestphis@supagro.inra.fr","guestphis","http://147.100.179.156:8080/phenomeapi/resources/")
aToken <- getToken("guestphis@supagro.inra.fr","guestphis")
aToken$data
}
}
