% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{setLogLevel}
\alias{setLogLevel}
\title{setLogLevel}
\usage{
setLogLevel(level = "INFO")
}
\arguments{
\item{level}{character, Default value values "INFO", Allowed values "CRITICAL"," ERROR", "WARNING", "INFO", "DEBUG", "NOTSET"}
}
\description{
Allows to retreive a particular level debugging messages
}
\seealso{
https://docs.python.org/3/library/logging.html#levels
}
