% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{showStatus}
\alias{showStatus}
\title{showStatus}
\usage{
showStatus(responseObject)
}
\arguments{
\item{responseObject}{objet de reponse HTTP httr}
}
\value{
character, message describes the response
}
\description{
Show status
}
\keyword{internal}
