% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connectToOpenSILEXWS}
\alias{connectToOpenSILEXWS}
\title{connectToOpenSILEXWS}
\usage{
connectToOpenSILEXWS(username, password, url)
}
\arguments{
\item{username}{character, login of the user to create the token}

\item{password}{character, password of the user to create the token}

\item{url}{character}
}
\description{
load name space and connexion parameters of the webservice.
Execute only once at the beginning of the requests.
In the case of a WebService change of address or a renaming of services, please edit this list.
and execute the function.
Demonstration instances:
\describe{
connectToOpenSILEXWS(username="guest@opensilex.org",
          password="guest", url = "http://www.opensilex.org/rest/")
}
}
