% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpCalls.R
\name{deleteResponseFromWS}
\alias{deleteResponseFromWS}
\title{deleteResponseFromWS}
\usage{
deleteResponseFromWS(resource, paramPath = NULL, attributes,
  encode = "json", requestBody)
}
\arguments{
\item{resource}{character, the name of the webservice resource}

\item{paramPath}{character, path URL encoded parameter}

\item{attributes}{query parameters}

\item{encode}{character, type of encodage}

\item{requestBody}{body data which will be send}
}
\value{
WSResponse WSResponse class instance
}
\description{
Create an URL to call the WS and return a formatted response of WSResponse class.
}
\keyword{internal}
