% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{retrieves a user identifier for connexion to the web service}
\usage{
getToken(identifier, password)
}
\arguments{
\item{identifier}{character, login of the user to create the token}

\item{password}{character, password of the user to create the token}
}
\value{
a session token user identifier in the WS
}
\description{
Retrieves a user identifier for connexion to the WebService (WS)
}
\keyword{internal}
