% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{retrieves a user identifier for connexion to the web service}
\usage{
getToken(login, password)
}
\arguments{
\item{login}{character, login of the user to create the token}

\item{password}{character, password of the user to create the token}
}
\value{
a session token user identifier in the WS
}
\description{
Retrieves a user identifier for connexion to the WebService (WS)
}
\examples{
\donttest{
connectToOpenSILEXWS(username="admin@opensilex.org",   password="admin", url = "http://138.102.159.37:8081/rest")

aToken$data
}
}
